godoc:vfs: extracted the `vfs` subpackage into an independent package
