+++ cmd/godoc/main.go
@@ -41,11 +41,11 @@ import (
 
 	"golang.org/x/tools/godoc"
 	"golang.org/x/tools/godoc/static"
-	"golang.org/x/tools/godoc/vfs"
-	"golang.org/x/tools/godoc/vfs/gatefs"
-	"golang.org/x/tools/godoc/vfs/mapfs"
-	"golang.org/x/tools/godoc/vfs/zipfs"
 	"golang.org/x/tools/internal/gocommand"
+	"golang.org/x/tools/vfs"
+	"golang.org/x/tools/vfs/gatefs"
+	"golang.org/x/tools/vfs/mapfs"
+	"golang.org/x/tools/vfs/zipfs"
 )
 
 const defaultAddr = "localhost:6060" // default webserver address
