+++ godoc/corpus.go
@@ -11,7 +11,7 @@ import (
 
 	"golang.org/x/tools/godoc/analysis"
 	"golang.org/x/tools/godoc/util"
-	"golang.org/x/tools/godoc/vfs"
+	"golang.org/x/tools/vfs"
 )
 
 // A Corpus holds all the state related to serving and indexing a
