+++ godoc/dirtrees_test.go
@@ -11,8 +11,8 @@ import (
 	"sort"
 	"testing"
 
-	"golang.org/x/tools/godoc/vfs"
-	"golang.org/x/tools/godoc/vfs/gatefs"
+	"golang.org/x/tools/vfs"
+	"golang.org/x/tools/vfs/gatefs"
 )
 
 func TestNewDirTree(t *testing.T) {
