+++ godoc/index.go
@@ -64,7 +64,7 @@ import (
 	"unicode"
 
 	"golang.org/x/tools/godoc/util"
-	"golang.org/x/tools/godoc/vfs"
+	"golang.org/x/tools/vfs"
 )
 
 // ----------------------------------------------------------------------------
