+++ godoc/pres.go
@@ -10,7 +10,7 @@ import (
 	"sync"
 	"text/template"
 
-	"golang.org/x/tools/godoc/vfs/httpfs"
+	"golang.org/x/tools/vfs/httpfs"
 )
 
 // SearchResultFunc functions return an HTML body for displaying search results.
