+++ godoc/server.go
@@ -28,7 +28,7 @@ import (
 
 	"golang.org/x/tools/godoc/analysis"
 	"golang.org/x/tools/godoc/util"
-	"golang.org/x/tools/godoc/vfs"
+	"golang.org/x/tools/vfs"
 )
 
 // handlerServer is a migration from an old godoc http Handler type.
