+++ godoc/template.go
@@ -38,7 +38,7 @@ import (
 	"regexp"
 	"strings"
 
-	"golang.org/x/tools/godoc/vfs"
+	"golang.org/x/tools/vfs"
 )
 
 // Functions in this file panic on error, but the panic is recovered
