+++ vfs/mapfs/mapfs_test.go
