+++ vfs/zipfs/zipfs_test.go
@@ -12,7 +12,7 @@ import (
 	"reflect"
 	"testing"
 
-	"golang.org/x/tools/godoc/vfs"
+	"golang.org/x/tools/vfs"
 )
 
 var (
